ft_strlen

+++ src/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 17:35:48 by irychkov          #+#    #+#             */
-/*   Updated: 2024/04/17 13:48:17 by irychkov         ###   ########.fr       */
+/*   Updated: 2024/04/29 17:46:03 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char	*str)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
